+++ se.gu.uoy.dsl/src/patterns/Trail.java
@@ -13,8 +13,8 @@ package patterns;
  * </p>
  * <ul>
  *   <li>{@link patterns.Trail#getObjetToFollow <em>Objet To Follow</em>}</li>
- *   <li>{@link patterns.Trail#getValue <em>Value</em>}</li>
  *   <li>{@link patterns.Trail#getRobots <em>Robots</em>}</li>
+ *   <li>{@link patterns.Trail#getValue <em>Value</em>}</li>
  * </ul>
  *
  * @see patterns.PatternsPackage#getTrail()
@@ -49,12 +49,12 @@ public interface Trail extends CompositePatterns, MissionDependent {
 	 * <!-- begin-user-doc -->
 	 * <!-- end-user-doc -->
 	 * @return the value of the '<em>Value</em>' attribute.
-	 * @see #setValue(int)
+	 * @see #setValue(double)
 	 * @see patterns.PatternsPackage#getTrail_Value()
 	 * @model required="true"
 	 * @generated
 	 */
-	int getValue();
+	double getValue();
 
 	/**
 	 * Sets the value of the '{@link patterns.Trail#getValue <em>Value</em>}' attribute.
@@ -64,7 +64,7 @@ public interface Trail extends CompositePatterns, MissionDependent {
 	 * @see #getValue()
 	 * @generated
 	 */
-	void setValue(int value);
+	void setValue(double value);
 
 	/**
 	 * Returns the value of the '<em><b>Robots</b></em>' reference.
